+++ src/Microsoft.Framework.Runtime.Roslyn/RoslynProjectReference.cs
@@ -1,38 +1,26 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Reflection.Metadata;
-using System.Reflection.PortableExecutable;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Emit;
-using Microsoft.CodeAnalysis.CSharp;
 
 namespace Microsoft.Framework.Runtime.Roslyn
 {
     public class RoslynProjectReference : IRoslynMetadataReference, IMetadataProjectReference
     {
-        private static readonly IList<IMetadataEmbeddedReference> _emptyList = new IMetadataEmbeddedReference[0];
-
-        private Lazy<IList<ResourceDescription>> _resources;
-        private bool _beforeCompileCalled;
-
         public RoslynProjectReference(CompilationContext compilationContext)
         {
             CompilationContext = compilationContext;
             MetadataReference = compilationContext.Compilation.ToMetadataReference(embedInteropTypes: compilationContext.Project.EmbedInteropTypes);
             Name = compilationContext.Project.Name;
-            _resources = new Lazy<IList<ResourceDescription>>(GetResources);
         }
 
-        public CompilationContext CompilationContext { get; set; }
-
-        public IList<ResourceDescription> Resources { get; set; }
+        public CompilationContext CompilationContext { get; private set; }
 
         public MetadataReference MetadataReference
         {
@@ -54,26 +42,8 @@ namespace Microsoft.Framework.Runtime.Roslyn
             }
         }
 
-        private void EnsureBeforeCompile()
-        {
-            if (_beforeCompileCalled)
-            {
-                return;
-            }
-            _beforeCompileCalled = true;
-            var context = new BeforeCompileContext(this);
-            foreach (var module in CompilationContext.Modules)
-            {
-                module.BeforeCompile(context);
-            }
-        }
-
-
-
         public IDiagnosticResult GetDiagnostics()
         {
-            EnsureBeforeCompile();
-
             var diagnostics = CompilationContext.Diagnostics
                 .Concat(CompilationContext.Compilation.GetDiagnostics());
 
@@ -82,8 +52,6 @@ namespace Microsoft.Framework.Runtime.Roslyn
 
         public IList<ISourceReference> GetSources()
         {
-            EnsureBeforeCompile();
-
             // REVIEW: Raw sources?
             return CompilationContext.Compilation
                                      .SyntaxTrees
@@ -95,12 +63,10 @@ namespace Microsoft.Framework.Runtime.Roslyn
 
         public Assembly Load(IAssemblyLoaderEngine loaderEngine)
         {
-            EnsureBeforeCompile();
-
             using (var pdbStream = new MemoryStream())
             using (var assemblyStream = new MemoryStream())
             {
-                IList<ResourceDescription> resources = _resources.Value;
+                IList<ResourceDescription> resources = CompilationContext.Resources;
 
                 Trace.TraceInformation("[{0}]: Emitting assembly for {1}", GetType().Name, Name);
 
@@ -152,16 +118,12 @@ namespace Microsoft.Framework.Runtime.Roslyn
 
         public void EmitReferenceAssembly(Stream stream)
         {
-            EnsureBeforeCompile();
-
             CompilationContext.Compilation.EmitMetadataOnly(stream);
         }
 
         public IDiagnosticResult EmitAssembly(string outputPath)
         {
-            EnsureBeforeCompile();
-
-            IList<ResourceDescription> resources = _resources.Value;
+            IList<ResourceDescription> resources = CompilationContext.Resources;
 
             var assemblyPath = Path.Combine(outputPath, Name + ".dll");
             var pdbPath = Path.Combine(outputPath, Name + ".pdb");
@@ -230,144 +192,6 @@ namespace Microsoft.Framework.Runtime.Roslyn
             }
         }
 
-        private IList<ResourceDescription> GetResources()
-        {
-            var resxProvider = new ResxResourceProvider();
-            var embeddedResourceProvider = new EmbeddedResourceProvider();
-
-            var resourceProvider = new CompositeResourceProvider(new IResourceProvider[] { resxProvider, embeddedResourceProvider });
-
-
-            var sw = Stopwatch.StartNew();
-            Trace.TraceInformation("[{0}]: Generating resources for {1}", GetType().Name, Name);
-
-            var resources = resourceProvider.GetResources(CompilationContext.Project);
-
-            sw.Stop();
-            Trace.TraceInformation("[{0}]: Generated resources for {1} in {2}ms", GetType().Name, Name, sw.ElapsedMilliseconds);
-
-            sw = Stopwatch.StartNew();
-            Trace.TraceInformation("[{0}]: Resolving required assembly neutral references for {1}", GetType().Name, Name);
-
-            var embeddedReferences = GetRequiredEmbeddedReferences();
-            resources.AddEmbeddedReferences(embeddedReferences);
-
-            Trace.TraceInformation("[{0}]: Resolved {1} required assembly neutral references for {2} in {3}ms",
-                GetType().Name,
-                embeddedReferences.Count,
-                Name,
-                sw.ElapsedMilliseconds);
-            sw.Stop();
-
-            return resources;
-        }
-
-        public IList<IMetadataEmbeddedReference> GetRequiredEmbeddedReferences()
-        {
-            var assemblyNeutralTypes = CompilationContext.MetadataReferences.OfType<IMetadataEmbeddedReference>()
-                                                         .ToDictionary(r => r.Name);
-
-            // No assembly neutral types so do nothing
-            if (assemblyNeutralTypes.Count == 0)
-            {
-                return _emptyList;
-            }
-
-            // Walk the assembly neutral references and embed anything that we use
-            // directly or indirectly
-            var results = GetUsedReferences(assemblyNeutralTypes);
-
-
-            // REVIEW: This should probably by driven by a property in the project metadata
-            if (results.Count == 0)
-            {
-                // If nothing outgoing from this assembly, treat it like a carrier assembly
-                // and embed everyting
-                foreach (var a in assemblyNeutralTypes.Keys)
-                {
-                    results.Add(a);
-                }
-            }
-
-            return results.Select(name => assemblyNeutralTypes[name])
-                          .ToList();
-        }
-
-        private HashSet<string> GetUsedReferences(Dictionary<string, IMetadataEmbeddedReference> assemblies)
-        {
-            var results = new HashSet<string>();
-
-            byte[] metadataBuffer = null;
-
-            // First we need to emit just the metadata for this compilation
-            using (var metadataStream = new MemoryStream())
-            {
-                var result = CompilationContext.Compilation.Emit(metadataStream);
-
-                if (!result.Success)
-                {
-                    // We failed to emit metadata so do nothing since we're probably
-                    // going to fail compilation anyways
-                    return results;
-                }
-
-                // Store the buffer and close the stream
-                metadataBuffer = metadataStream.ToArray();
-            }
-
-            var stack = new Stack<Tuple<string, byte[]>>();
-            stack.Push(Tuple.Create((string)null, metadataBuffer));
-
-            while (stack.Count > 0)
-            {
-                var top = stack.Pop();
-
-                var assemblyName = top.Item1;
-
-                if (!string.IsNullOrEmpty(assemblyName) &&
-                    !results.Add(assemblyName))
-                {
-                    // Skip the reference if saw it already
-                    continue;
-                }
-
-                var buffer = top.Item2;
-
-                foreach (var reference in GetReferences(buffer))
-                {
-                    IMetadataEmbeddedReference embeddedReference;
-                    if (assemblies.TryGetValue(reference, out embeddedReference))
-                    {
-                        stack.Push(Tuple.Create(reference, embeddedReference.Contents));
-                    }
-                }
-            }
-
-            return results;
-        }
-
-        private static IList<string> GetReferences(byte[] buffer)
-        {
-            var references = new List<string>();
-
-            using (var stream = new MemoryStream(buffer))
-            {
-                var peReader = new PEReader(stream);
-
-                var reader = peReader.GetMetadataReader();
-
-                foreach (var a in reader.AssemblyReferences)
-                {
-                    var reference = reader.GetAssemblyReference(a);
-                    var referenceName = reader.GetString(reference.Name);
-
-                    references.Add(referenceName);
-                }
-
-                return references;
-            }
-        }
-
         private static DiagnosticResult CreateDiagnosticResult(bool success, IEnumerable<Diagnostic> diagnostics)
         {
             var formatter = new DiagnosticFormatter();
@@ -379,44 +203,6 @@ namespace Microsoft.Framework.Runtime.Roslyn
                                   .Select(d => formatter.Format(d)).ToList();
 
             return new DiagnosticResult(success, warnings, errors);
-        }
-
-        private class BeforeCompileContext : IBeforeCompileContext
-        {
-            private readonly RoslynProjectReference _roslynProjectReference;
-
-            public BeforeCompileContext(RoslynProjectReference roslynProjectReference)
-            {
-                _roslynProjectReference = roslynProjectReference;
-            }
-
-            public CSharpCompilation CSharpCompilation
-            {
-                get
-                {
-                    return _roslynProjectReference.CompilationContext.Compilation;
-                }
-                set
-                {
-                    _roslynProjectReference.CompilationContext.Compilation = value;
-                }
-            }
-
-            public IList<Diagnostic> Diagnostics
-            {
-                get
-                {
-                    return _roslynProjectReference.CompilationContext.Diagnostics;
-                }
-            }
-
-            public IList<ResourceDescription> Resources
-            {
-                get
-                {
-                    return _roslynProjectReference._resources.Value;
-                }
-            }
         }
     }
 }
